refactor: move GsonConfigurationCodec to example module, fix init error handling
- Move GsonConfigurationCodec to example module (it's a demo/example codec,
  not part of the framework public API)
- Update CustomClientActivity to use GsonConfigurationCodec + FileBackedConfigStore
  to demonstrate human-readable JSON cache
- Fix initialization error handling in AndroidBaseClient and EppoClient:
  complete exceptionally when ex != null, avoiding double increment of failCount
